+++ build.gradle.kts
@@ -58,9 +58,6 @@ dependencies {
 	runtimeOnly("io.jsonwebtoken:jjwt-impl:0.11.5")
 	runtimeOnly("io.jsonwebtoken:jjwt-jackson:0.11.5")
 
-	// BCrypt
-	implementation("org.springframework.boot:spring-boot-starter-security")
-
 	// firebase auth
 //	implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
 	implementation("org.springframework.boot:spring-boot-starter-security")
