--- src/main/kotlin/com/Dnight/calinify/auth/BCryptConfig.kt
@@ -1,13 +1,0 @@
-package com.dnight.calinify.auth
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-import org.springframework.security.crypto.password.PasswordEncoder
-
-@Configuration
-class BCryptConfig {
-
-    @Bean
-    fun passwordEncoder(): PasswordEncoder = BCryptPasswordEncoder()
-}
