--- src/main/kotlin/com/Dnight/calinify/auth/TokenInfo.kt
@@ -1,6 +1,0 @@
-package com.dnight.calinify.auth
-
-data class TokenInfo(
-    val grantType: String,
-    val accessToken: String,
-)
