+++ src/main/kotlin/com/Dnight/calinify/auth/dto/response/TokenResponseDTO.kt
@@ -1,0 +1,7 @@
+package com.dnight.calinify.auth.dto.response
+
+data class TokenResponseDTO(
+    val tokenType: String? = "bearer",
+    val accessToken: String,
+    val refreshToken: String
+)
